Phase8/requirements.txt
mcp
requests
fastapi
strands-agents
bedrock-agentcore[strands-agents]
streamlit